ADD,CHG: BICenhanced.setEnforceFlatWeightsVoting() because:
BICenhanced must have weights -- they are used in the fail-safe
label insertion code (to choose suitable input label), and the voting
is considering whatever is the content of the inWeights vector
(so far the WeightedVotingLabelFuserWithFailSafe fuser was always used).

If, however, flat voting is demanded, we still have to be provided
with valid (possibly not flat) inWeights vector but have to switch
to use the ForcedFlatVotingLabelFuserWithFailSafe fuser.